Cast IndType to int before comparison.

+++ inc/griddingFunctions.hpp
@@ -146,11 +146,11 @@ __inline__ __device__ __host__ bool isOutlier(int x, int y, int z, int center_x,
 
 __inline__ __device__ __host__ bool isOutlier(int x, int y, int z, int center_x, int center_y, int center_z, IndType3 dim, int sector_offset)
 {
-  return ((center_x - sector_offset + x) >= dim.x ||
+  return ((center_x - sector_offset + x) >= (int)dim.x ||
     (center_x - sector_offset + x) < 0 ||
-    (center_y - sector_offset + y) >= dim.y ||
+    (center_y - sector_offset + y) >= (int)dim.y ||
     (center_y - sector_offset + y) < 0 ||
-    (center_z - sector_offset + z) >= dim.z ||
+    (center_z - sector_offset + z) >= (int)dim.z ||
     (center_z - sector_offset + z) < 0);
 }
 
@@ -165,9 +165,9 @@ __inline__ __device__ __host__ bool isOutlier2D(int x, int y, int center_x, int 
 
 __inline__ __device__ __host__ bool isOutlier2D(int x, int y, int center_x, int center_y, IndType3 dim, int sector_offset)
 {
-  return ((center_x - sector_offset + x) >= dim.x ||
+  return ((center_x - sector_offset + x) >= (int)dim.x ||
     (center_x - sector_offset + x) < 0 ||
-    (center_y - sector_offset + y) >= dim.y ||
+    (center_y - sector_offset + y) >= (int)dim.y ||
     (center_y - sector_offset + y) < 0);
 }
